Changed some of the datastructures and made the code more efficient by removing redundancies.
There are still some weird stuff where the pointer references are a bit over the top because
golang does some of these things automatically but I'm not sure about them.

Final testing, either I use this as a binding for python or would be fun to construct an svm from scratch.